requirement.txt
langchain
python-dotenv
langchain_community
pypdf
beautifulsoup4
arxiv
PyMuPDF
wikipedia
langchain-text-splitters
pydantic
langchain-groq
langchain-core
langchain-community
langchain_openai
defusedxml
langchain_huggingface
faiss-cpu
langchain_google_genai
langchain-pinecone
uuid
sentence-transformers
python-docx
youtube_search
langchain_mistral